refactor : result 클래스들 result 패키지로 이동
