+++ src/main/java/model/result/ParticipantWinningResult.java
@@ -1,4 +1,4 @@
-package model;
+package model.result;
 
 import java.util.Collections;
 import java.util.HashMap;
